+++ poetry/console/commands/run.py
@@ -11,10 +11,46 @@ class RunCommand(VenvCommand):
 
     def handle(self):
         args = self.argument('args')
+        script = args[0]
+        scripts = self.poetry.local_config.get('scripts')
+
+        if scripts and script in scripts:
+            return self.run_script(scripts[script], args)
 
         venv = self.venv
 
         return venv.execute(*args)
+
+    def run_script(self, script, args):
+        module, callable_ = script.split(':')
+
+        src_in_sys_path = 'sys.path.append(\'src\'); ' \
+            if self._module.is_in_src() else ''
+
+        cmd = ['python', '-c']
+
+        cmd += [
+            '"import sys; '
+            'from importlib import import_module; '
+            'sys.argv = {!r}; {}'
+            'import_module(\'{}\').{}()"'.format(
+                args, src_in_sys_path, module, callable_
+            )
+        ]
+
+        return self.venv.run(*cmd, shell=True, call=True)
+
+    @property
+    def _module(self):
+        from ...masonry.utils.module import Module
+
+        poetry = self.poetry
+        package = poetry.package
+        path = poetry.file.parent
+        module = Module(
+            package.name, path.as_posix()
+        )
+        return module
 
     def merge_application_definition(self, merge_args=True):
         if self._application is None \